cleanup implementation, docs
