--- Chakra.Core.Async/Data/IDataSession.cs
@@ -1,32 +1,0 @@
-﻿using System;
-using ZenProgramming.Chakra.Core.Async.Data.Repositories;
-using ZenProgramming.Chakra.Core.Data.Repositories;
-
-namespace ZenProgramming.Chakra.Core.Data
-{
-    ///// <summary>
-    ///// Represents inteface for session of data
-    ///// </summary>
-    //public interface IDataSessionAsync : IDataSession
-    //{
-    //    ///// <summary>
-    //    ///// Begin new transaction on active session
-    //    ///// </summary>
-    //    ///// <returns>Returns data transaction instance</returns>
-    //    //IDataTransactionAsync BeginTransactionAsync();
-        
-    //    ///// <summary>
-    //    ///// Execute resolve of repository interface on concrete
-    //    ///// type that matchh data provider defined by data session
-    //    ///// </summary>
-    //    ///// <typeparam name="TRepositoryInterface">Type of repository interface</typeparam>
-    //    ///// <returns>Returns repository instance</returns>
-    //    //TRepositoryInterface ResolveRepositoryAsync<TRepositoryInterface>()
-    //    //    where TRepositoryInterface : IRepositoryAsync;
-
-    //    ///// <summary>
-    //    ///// Active transaction on data session
-    //    ///// </summary>
-    //    //IDataTransactionAsync TransactionAsync { get; }
-    //}
-}
