+++ Chakra.Core.EntityFramework/Data/EntityFrameworkDataSession.cs
@@ -37,7 +37,7 @@ namespace ZenProgramming.Chakra.Core.EntityFramework.Data
         /// Set active transaction on current instance
         /// </summary>
         /// <param name="transaction">Transaction instance</param>
-        public void SetActiveTransaction(EntityFrameworkDataTransaction<TDbContext> transaction)
+        public void SetActiveTransaction(IEntityFrameworkTransaction<TDbContext> transaction)
         {
             //Imposto la transazione
             Transaction = transaction;
@@ -69,13 +69,13 @@ namespace ZenProgramming.Chakra.Core.EntityFramework.Data
         /// <summary>
         /// Active transaction on data session
         /// </summary>
-        public IDataTransaction Transaction { get; private set; }
+        public IDataTransaction Transaction { get; protected set; }
 
         /// <summary>
         /// Begin new transaction on active session
         /// </summary>
         /// <returns>Returns data transaction instance</returns>
-        public IDataTransaction BeginTransaction()
+        public virtual IDataTransaction BeginTransaction()
         {
             //Ritorno un'istanza della transazone di Entity Framework
             return new EntityFrameworkDataTransaction<TDbContext>(this);
