--- Chakra.Core.Mocks.Async/Data/MockDataSessionAsync.cs
@@ -1,14 +1,0 @@
-﻿using ZenProgramming.Chakra.Core.Mocks.Scenarios;
-using ZenProgramming.Chakra.Core.Mocks.Scenarios.Options;
-
-namespace ZenProgramming.Chakra.Core.Mocks.Async.Data
-{
-    ///// <summary>
-    ///// Simplified implementation of data session on "Mock" engine
-    ///// </summary>
-    ///// <typeparam name="TScenarioInstance">Type of scenario instance</typeparam>
-    //public class MockDataSessionAsync<TScenarioInstance> : MockDataSessionAsync<TScenarioInstance, ScopedScenarioOption<TScenarioInstance>>
-    //    where TScenarioInstance : class, IScenario, new()
-    //{ 
-    //}
-}
