--- Chakra.Core.Mocks.Async/Data/MockDataSessionAsync.generic.cs
@@ -1,30 +1,0 @@
-﻿using System;
-using ZenProgramming.Chakra.Core.Async.Data.Repositories;
-using ZenProgramming.Chakra.Core.Data;
-using ZenProgramming.Chakra.Core.Mocks.Data;
-using ZenProgramming.Chakra.Core.Mocks.Data.Repositories;
-using ZenProgramming.Chakra.Core.Mocks.Scenarios;
-using ZenProgramming.Chakra.Core.Mocks.Scenarios.Options;
-
-namespace ZenProgramming.Chakra.Core.Mocks.Async.Data
-{
-  //  /// <summary>
-  //  /// Generic implementation of data session on "Mock" engine with injectable scenario option
-  //  /// </summary>
-  //  /// <typeparam name="TScenarioImplementation">Type of scenario instance (ex: SimpleScenario)</typeparam>
-  //  /// <typeparam name="TScenarioOption">Type of scenario option (ex: TransientScenarioOption[SimpleScenario])</typeparam>
-  //  public class MockDataSessionAsync<TScenarioImplementation, TScenarioOption> : MockDataSession<TScenarioImplementation, TScenarioOption>, IMockDataSessionAsync
-  //      where TScenarioImplementation : class, IScenario, new()
-  //      where TScenarioOption : class, IScenarioOption<TScenarioImplementation>, new()
-  //  {
-
-  //      /// <summary>
-  //      /// Finalizer that ensures the object is correctly disposed of.
-		///// </summary>
-  //      ~MockDataSessionAsync()
-  //      {
-  //          //Richiamo i dispose implicito
-  //          Dispose(false);
-  //      }
-  //  }
-}
