+++ Chakra.Core.Mocks.Async/Data/MockDataTransactionAsync.cs
@@ -1,88 +1,49 @@
 ﻿using System;
 using System.Threading.Tasks;
+using ZenProgramming.Chakra.Core.Async.Data;
 using ZenProgramming.Chakra.Core.Data;
-using ZenProgramming.Chakra.Core.Data.Repositories;
-using ZenProgramming.Chakra.Core.Data.Repositories.Helpers;
+using ZenProgramming.Chakra.Core.Diagnostic;
+using ZenProgramming.Chakra.Core.Mocks.Data;
 
 namespace ZenProgramming.Chakra.Core.Mocks.Async.Data
 {
-    public static class MockDataTransactionAsyncExtension
+    /// <summary>
+    /// Data transaction implementation for mockup
+    /// </summary>
+    public class MockDataTransactionAsync : MockDataTransaction, IDataTransactionAsync
     {
-        // TIPS: customizzare classe dataTransaction
-        //public static Task CommitAsync(this IDataTransaction dataTransaction)
-        //{
-        //    if (dataTransaction is IDataTransaction _dt)
-        //    {
-        //        return _dt.CommitAsync();
-        //    }
-        //    throw new InvalidCastException($"Unable to cast type of '{dataTransaction.GetType().FullName}' to " +
-        //                                   $"interface '{typeof(IDataTransaction).FullName}'.");
-        //}
-
-        //public static Task RollBackAsync(this IDataTransaction dataTransaction)
-        //{
-        //    if (dataTransaction is IDataTransaction _dt)
-        //    {
-        //        return _dt.RollBackAsync();
-        //    }
-        //    throw new InvalidCastException($"Unable to cast type of '{dataTransaction.GetType().FullName}' to " +
-        //                                   $"interface '{typeof(IDataTransaction).FullName}'.");
-        //}
-
-        public static Task CommitAsync(this IDataTransaction mockDataTransaction)
+        /// <summary>
+        /// Constructor
+        /// </summary>
+        public MockDataTransactionAsync(IMockDataSession dataSession): base(dataSession)
         {
-            // Call base method
-            mockDataTransaction.Commit();
+        }
+
+        /// <summary>
+        /// Executes commit async of transaction
+        /// </summary>
+        public Task CommitAsync()
+        {
+            this.Commit();
             return Task.CompletedTask;
         }
 
-        public static Task RollBackAsync(this IDataTransaction mockDataTransaction)
+        /// <summary>
+        /// Executes rollback of transaction
+        /// </summary>
+        public Task RollBackAsync()
         {
-            // Call base method
-            mockDataTransaction.Rollback();
+            this.Rollback();
             return Task.CompletedTask;
+        }
+
+        /// <summary>
+        /// Finalizes that ensures the object is correctly disposed of.
+		/// </summary>
+        ~MockDataTransactionAsync()
+        {
+            //Implicit dispose
+            Dispose(false);
         }
     }
-
-  //  /// TIPS: customizzare classe dataTransaction
-  //  /// <summary>
-  //  /// Data transaction implementation for mockup
-  //  /// </summary>
-  //  public class MockDataTransactionAsync : MockDataTransaction, IDataTransactionAsync
-  //  {
-  //      /// <summary>
-  //      /// Constructor
-  //      /// </summary>
-  //      public MockDataTransactionAsync(IMockDataSessionAsync dataSession): base(dataSession) {}
-
-  //      /// <summary>
-  //      /// Executes commit of transaction
-  //      /// </summary>
-  //      public Task CommitAsync()
-  //      {
-  //          // Call base method
-  //          this.Commit();
-  //          return Task.CompletedTask;
-  //      }
-
-  //      /// <summary>
-  //      /// Executes rollback of transaction
-  //      /// </summary>
-  //      public Task RollBackAsync()
-  //      {
-  //          // Call base method
-  //          this.Rollback();
-  //          return Task.CompletedTask;
-  //      }
-
-  //      /// <summary>
-  //      /// Finalizes that ensures the object is correctly disposed of.
-		///// </summary>
-  //      ~MockDataTransactionAsync()
-  //      {
-  //          //Implicit dispose
-  //          Dispose(false);
-  //      }
-
-  //  }
 }
