+++ Chakra.Core.Mocks.Async/Data/Repositories/IMockRepositoryAsync.cs
@@ -3,7 +3,7 @@
 namespace ZenProgramming.Chakra.Core.Mocks.Data.Repositories
 {
     /// <summary>
-    /// Public interface for mock repositories (used for auto-discovery!)
+    /// Public interface for mock repositories async (used for auto-discovery!)
     /// </summary>
     public interface IMockRepositoryAsync : IRepositoryAsync
     {
