+++ Chakra.Core.Mocks.Tests/MockCarRepositoryTests.cs
@@ -2,10 +2,10 @@
 using System.Linq;
 using System.Threading.Tasks;
 using Xunit;
+using ZenProgramming.Chakra.Core.Async.Data;
 using ZenProgramming.Chakra.Core.Data;
 using ZenProgramming.Chakra.Core.Mocks.Async.Data;
 using ZenProgramming.Chakra.Core.Mocks.Async.Data.Extensions;
-using ZenProgramming.Chakra.Core.Mocks.Async.Scenarios.Extensions;
 using ZenProgramming.Chakra.Core.Mocks.Data;
 using ZenProgramming.Chakra.Core.Mocks.Scenarios.Extensions;
 using ZenProgramming.Chakra.Core.Mocks.Scenarios.Options;
@@ -22,7 +22,7 @@ namespace ZenProgramming.Chakra.Core.Mocks.Tests
         public MockCarRepositoryTests() 
         {
             //Register default session and open
-            _DataSession = SessionFactory.OpenSession<MockDataSession<
+            _DataSession = SessionFactory.OpenSession<MockDataSessionAsync<
                 SimpleScenario,
                 TransientScenarioOption<SimpleScenario>>>();
         }
