+++ Chakra.Core.Mocks.Tests/MockPersonRepositoryTests.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Linq;
-using System.Threading.Tasks;
 using Xunit;
 using ZenProgramming.Chakra.Core.Data;
 using ZenProgramming.Chakra.Core.Mocks.Data;
@@ -9,7 +8,7 @@ using ZenProgramming.Chakra.Core.Mocks.Scenarios.Options;
 using ZenProgramming.Chakra.Core.Tests.Environment.Entities;
 using ZenProgramming.Chakra.Core.Tests.Environment.Repositories;
 using ZenProgramming.Chakra.Core.Mocks.Tests.Environment.Scenarios;
-using ZenProgramming.Chakra.Core.Mocks.Async.Data.Repositories;
+
 namespace ZenProgramming.Chakra.Core.Mocks.Tests
 {
     public class MockPersonRepositoryTests : IDisposable
