+++ Chakra.Core.MongoDb.Async/Data/Repositories/MongoDbRepositoryAsyncExtensions.Async.cs
@@ -1,14 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
 using MongoDB.Driver;
 using MongoDB.Driver.Linq;
 using ZenProgramming.Chakra.Core.Async.Data.Repositories;
 using ZenProgramming.Chakra.Core.Data;
-using ZenProgramming.Chakra.Core.Data.Repositories;
-using ZenProgramming.Chakra.Core.Data.Repositories.Helpers;
 using ZenProgramming.Chakra.Core.Entities;
 using ZenProgramming.Chakra.Core.MongoDb.Data.Options;
