+++ Chakra.Core.MongoDb/Data/Repositories/MongoDbRepositoryBase.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Linq.Expressions;
-using System.Threading.Tasks;
 using MongoDB.Driver;
 using MongoDB.Driver.Linq;
 using ZenProgramming.Chakra.Core.Data;
