+++ Chakra.Core.Tests/Environment/Repositories/IDepartmentRepository.cs
@@ -1,5 +1,4 @@
-﻿using ZenProgramming.Chakra.Core.Async.Data.Repositories;
-using ZenProgramming.Chakra.Core.Data.Repositories;
+﻿using ZenProgramming.Chakra.Core.Data.Repositories;
 using ZenProgramming.Chakra.Core.Tests.Environment.Entities;
 
 namespace ZenProgramming.Chakra.Core.Tests.Environment.Repositories
