+++ Chakra.Core.WebApi/Filters/Models/ResponseTrace.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using ZenProgramming.Chakra.Core.WebApi.Filters.Models;
 
 namespace ZenProgramming.Chakra.Core.WebApi.Filters.Models
 {
