+++ Chakra.Core.WebApi/Filters/TraceErrorOnFileAttribute.cs
@@ -2,7 +2,6 @@
 using System.IO;
 using ZenProgramming.Chakra.Core.WebApi.Filters.Helpers;
 using ZenProgramming.Chakra.Core.WebApi.Filters.Models;
-using ZenProgramming.Chakra.Core.WebApi.Filters;
 
 namespace ZenProgramming.Chakra.Core.WebApi.Filters
 {
