+++ Chakra.Core.WebApi/Helpers/ClientDetections/ClientCapabilities.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using ZenProgramming.Chakra.Core.WebApi.Helpers.ClientDetections;
 
 namespace ZenProgramming.Chakra.Core.WebApi.Helpers.ClientDetections
 {
